requirements.txt
# Web框架
fastapi>=0.104.1
uvicorn[standard]>=0.24.0

# 文件处理
python-multipart>=0.0.6
Pillow>=10.1.0

# HTTP客户端
requests>=2.31.0

# 安全相关
cryptography>=41.0.7

# 开发工具
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.1

# 环境变量
python-dotenv>=1.0.0
